feat: #526 organize node modules utils
