--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/alias/cache.json
@@ -1,1 +1,0 @@
-{}
