--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/alias/invalid.json
@@ -1,1 +1,0 @@
-This is not valid JSON!
