--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/alias/multiple.json
@@ -1,10 +1,0 @@
-{
-	"unpkg": {
-		"fs": "./fs-shim2.js",
-		"./log": "./log-shim.js"
-	},
-	"browser": {
-		"fs": "./fs-shim.js",
-		"./random.js": "./random-shim.js"
-	}
-}
