--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/alias/normalize.json
@@ -1,5 +1,0 @@
-{
-	"browser": {
-		"/random.js": "./random-shim.js"
-	}
-}
