--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/packages/a-package/package.json
@@ -1,4 +1,0 @@
-{
-	"name": "a-package",
-	"version": "1.0.0"
-}
