--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/packages/a-package/with-package-json/no-dot/with-extension/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "file.js"
-}
