--- packages/liferay-js-toolkit-core/src/__tests__/__fixtures__/packages/a-package/with-package-json/with-dot/no-extension/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "./file"
-}
