+++ packages/liferay-js-toolkit-core/src/node/__tests__/modules.test.ts
@@ -13,38 +13,6 @@ it('isLocalModule() works', () => {
 	expect(mod.isLocalModule('a-module')).toBe(false);
 });
 
-it('isNodeCoreModule() works', () => {
-	expect(mod.isNodeCoreModule('fs')).toBe(true);
-	expect(mod.isNodeCoreModule('path')).toBe(true);
-	expect(mod.isNodeCoreModule('process')).toBe(true);
-	expect(mod.isNodeCoreModule('http')).toBe(true);
-
-	expect(mod.isNodeCoreModule('./a-module')).toBe(false);
-	expect(mod.isNodeCoreModule('../a-module')).toBe(false);
-	expect(mod.isNodeCoreModule('/a-module')).toBe(false);
-	expect(mod.isNodeCoreModule('a-module')).toBe(false);
-});
-
-it('isReservedDependency() works', () => {
-	expect(mod.isReservedDependency('require')).toBe(true);
-	expect(mod.isReservedDependency('module')).toBe(true);
-	expect(mod.isReservedDependency('exports')).toBe(true);
-
-	expect(mod.isReservedDependency('./a-module')).toBe(false);
-	expect(mod.isReservedDependency('../a-module')).toBe(false);
-	expect(mod.isReservedDependency('/a-module')).toBe(false);
-	expect(mod.isReservedDependency('a-module')).toBe(false);
-});
-
-it('isExternalDependency() works', () => {
-	expect(mod.isExternalDependency('a-module')).toBe(true);
-	expect(mod.isExternalDependency('fs')).toBe(true);
-
-	expect(mod.isExternalDependency('./a-module')).toBe(false);
-	expect(mod.isExternalDependency('../a-module')).toBe(false);
-	expect(mod.isExternalDependency('/a-module')).toBe(false);
-});
-
 it('joinModuleName() works', () => {
 	expect(mod.joinModuleName('@a-scope', 'a-package', '/a-dir/a-module')).toBe(
 		'@a-scope/a-package/a-dir/a-module'
