+++ packages/liferay-js-toolkit-core/src/node/namespace.ts
